refactor(adblocker): reduce the scriptlet-injection change surface
Keep the registration lifecycle and injection code close to their previous
structure, fold the guard wrapper into the codegen module, and drop
unrelated cleanups so the branch stays focused on scriptlet injection.
