+++ .gitignore
@@ -44,4 +44,3 @@ xcuserdata
 
 # WDIO
 .wdio
-.claude/
