+++ eslint.config.js
@@ -48,13 +48,7 @@ export default [
     },
   },
   {
-    ignores: [
-      '.wdio',
-      'dist/*',
-      'xcode/*',
-      'src/rule_resources*',
-      '**/*.generated.js',
-      'web-ext-artifacts/*',
-    ],
+    ignores: ['.wdio', 'dist/*', 'xcode/*', 'src/rule_resources*', 'web-ext-artifacts/*'],
   },
+  { ignores: ['**/*.generated.js'] },
 ];
