+++ src/background/serp.js
@@ -75,7 +75,7 @@ chrome.webNavigation.onCommitted.addListener(async (details) => {
       chrome.scripting.executeScript(
         {
           injectImmediately: true,
-          world: 'ISOLATED',
+          world: chrome.scripting.ExecutionWorld?.ISOLATED ?? 'ISOLATED',
           target: {
             tabId: details.tabId,
           },
